+++ shop/src/main/java/simpleshop/shop/domain/ItemForm.java
@@ -5,10 +5,10 @@ import lombok.Data;
 @Data
 public class ItemForm {
     private String itemName;
-    private Integer price;
+    private double price;
     private Integer quantity;
 
-    public ItemForm(String itemName, Integer price, Integer quantity) {
+    public ItemForm(String itemName, double price, Integer quantity) {
         this.itemName = itemName;
         this.price = price;
         this.quantity = quantity;
