+++ shop/src/main/java/simpleshop/shop/service/ItemService.java
@@ -10,7 +10,7 @@ public interface ItemService {
     Item getItemById(UUID id);
     List<Item> getAllItems();
     //등록
-    Item addItem(String itemName, Integer price, Integer quantity);
+    Item addItem(String itemName, double price, Integer quantity);
     //수정
     Item update(UUID id, Item item);
     //삭제
